requirements.txt
# Core ML / DL
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.35.0
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0

# NLP
nltk>=3.8.0
tokenizers>=0.15.0

# Evaluation & Metrics
scipy>=1.11.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Experiment tracking
tqdm>=4.65.0
pyyaml>=6.0

# Data download
requests>=2.31.0
datasets>=2.14.0   # HuggingFace datasets (for news/rec datasets)

# Notebook
jupyter>=1.0.0
ipykernel>=6.0.0
streamlit>=1.35.0

# Linting (optional CI)
flake8>=6.0.0
